requirements.txt
discord.py==2.0.0
cryptography==39.0.1
python-dotenv==0.20.0
